Impl SVG schema in the schemas package
