+++ packages/generator/package.json
