+++ packages/schemas/package.json
@@ -45,6 +45,7 @@
   "dependencies": {
     "@pdfme/pdf-lib": "^1.17.3",
     "bwip-js": "^4.1.1",
+    "fast-xml-parser": "^4.3.2",
     "fontkit": "^2.0.2"
   },
   "devDependencies": {
