+++ packages/schemas/src/index.ts
@@ -1,5 +1,6 @@
 import text from './text/index.js';
 import image from './graphics/image.js';
+import svg from './graphics/svg.js';
 import barcodes from './barcodes/index.js';
 import line from './shapes/line.js';
 import { rectangle, ellipse } from './shapes/rectAndEllipse.js';
@@ -10,6 +11,7 @@ const builtInPlugins = { Text: text };
 export {
   text,
   image,
+  svg,
   barcodes,
   line,
   rectangle,
