+++ playground/src/plugins/index.ts
@@ -1,9 +1,7 @@
 import { signature } from './signature';
-import { svg } from './svg';
 
 const plugins = {
   signature,
-  svg,
 };
 
 export default plugins;
